Struct porteurbars::Template

pub struct Template {
    pub path: PathBuf,
}

A template holds a path to template source and a file describing the default values associated with names used in the template

Fields

path

path to template source

Methods

impl Template

fn validate(path: &Path) -> Result<bool>

validates a template located at path

fn init(force: bool) -> Result<()>

initializes current working directory with porteurbar defaults

fn download(repo: &str, tag: Option<&str>) -> Result<bool>

downloads a template from repo (user/repo) todo: handle host (ghe), branch, credentials (private repo)

fn list() -> Result<Vec<String>>

fn delete(tag: &str) -> Result<bool>

fn get<P>(path: P) -> Result<Template> where P: AsRef<Path>

Resolve template

fn apply(&self, target: &Path) -> Result<()>

Apply template